+++ alicompat/CHANGELOG.md
@@ -1,3 +1,8 @@
+## []
+
+- Added GregTech CEu Modern support
+- Added Kaleidoscope Cookery support
+
 ## [1.0.0]
 
 - Initial release